// src/app/branding-sponsorship/model/branding-sponsorship.model.ts
export class BrandingSponsorshipModel {

    _001:  BrandingSponsorshipDetailsModel;
    _002:  BrandingSponsorshipDetailsModel;
    _003:  BrandingSponsorshipDetailsModel;
    _004:  BrandingSponsorshipDetailsModel;
    _005:  BrandingSponsorshipDetailsModel;
    _005a: BrandingSponsorshipDetailsModel;
    _006:  BrandingSponsorshipDetailsModel;
    _007:  BrandingSponsorshipDetailsModel;
    _008:  BrandingSponsorshipDetailsModel;
    _009:  BrandingSponsorshipDetailsModel;
    _010:  BrandingSponsorshipDetailsModel;
    _011:  BrandingSponsorshipDetailsModel;
    _012:  BrandingSponsorshipDetailsModel;
    _014:  BrandingSponsorshipDetailsModel;
    _013:  BrandingSponsorshipDetailsModel;
    _015:  BrandingSponsorshipDetailsModel;
    _016:  BrandingSponsorshipDetailsModel;
    _017:  BrandingSponsorshipDetailsModel;
    _018:  BrandingSponsorshipDetailsModel;
    _019:  BrandingSponsorshipDetailsModel;
    _020:  BrandingSponsorshipDetailsModel;
    _021:  BrandingSponsorshipDetailsModel;
    _022:  BrandingSponsorshipDetailsModel;
    _023:  BrandingSponsorshipDetailsModel;
    _024:  BrandingSponsorshipDetailsModel;
    _025:  BrandingSponsorshipDetailsModel;
    _026:  BrandingSponsorshipDetailsModel;
    _027:  BrandingSponsorshipDetailsModel;
    _028:  BrandingSponsorshipDetailsModel;
    _029:  BrandingSponsorshipDetailsModel;
    _030:  BrandingSponsorshipDetailsModel;

    constructor() { 
		this._001  =  new BrandingSponsorshipDetailsModel();
		this._002  =  new BrandingSponsorshipDetailsModel();
		this._003  =  new BrandingSponsorshipDetailsModel();
		this._004  =  new BrandingSponsorshipDetailsModel();
		this._005  =  new BrandingSponsorshipDetailsModel();
		this._005a =  new BrandingSponsorshipDetailsModel();
		this._006  =  new BrandingSponsorshipDetailsModel();
		this._007  =  new BrandingSponsorshipDetailsModel();
		this._008  =  new BrandingSponsorshipDetailsModel();
		this._009  =  new BrandingSponsorshipDetailsModel();
		this._010  =  new BrandingSponsorshipDetailsModel();
		this._011  =  new BrandingSponsorshipDetailsModel();
		this._012  =  new BrandingSponsorshipDetailsModel();
		this._014  =  new BrandingSponsorshipDetailsModel();
		this._013  =  new BrandingSponsorshipDetailsModel();
		this._015  =  new BrandingSponsorshipDetailsModel();
		this._016  =  new BrandingSponsorshipDetailsModel();
		this._017  =  new BrandingSponsorshipDetailsModel();
		this._018  =  new BrandingSponsorshipDetailsModel();
		this._019  =  new BrandingSponsorshipDetailsModel();
		this._020  =  new BrandingSponsorshipDetailsModel();
		this._021  =  new BrandingSponsorshipDetailsModel();
		this._022  =  new BrandingSponsorshipDetailsModel();
		this._023  =  new BrandingSponsorshipDetailsModel();
		this._024  =  new BrandingSponsorshipDetailsModel();
		this._025  =  new BrandingSponsorshipDetailsModel();
		this._026  =  new BrandingSponsorshipDetailsModel();
		this._027  =  new BrandingSponsorshipDetailsModel();
		this._028  =  new BrandingSponsorshipDetailsModel();
		this._029  =  new BrandingSponsorshipDetailsModel();
		this._030  =  new BrandingSponsorshipDetailsModel();
	}

}

export class BrandingSponsorshipDetailsModel {

    slNo: string | number = '';
    checked: boolean = false;
    size: string = '';
    location: string | number = '';
    type: string = '';
    code: string = '';
    rate: string = '';
    active: string = '';

    constructor() { }
    
}